File saving and loading fixed
